Move Wordlestat functions to their own file
